+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 - Fetch prompts from API instead of github repo
 - Fix display of prompt details
+- Allow configure if show detailed view in prompts list via preferences
 
 ## [Initial Version] - 2024-09-10